refactored MetricTemplateVersionCompare and MetricVersonCompare components to use HOC
